Extract KeyLockManager to separate lock_manager.rs file
Move KeyLockManager, KeyLockEntry, MultiKeyGuard and their implementations
along with tests to a dedicated lock_manager.rs module.
